pyproject.toml
# pyproject.toml

[build-system]
requires      = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "airflux"
version = "1.0.0"
description = "A simple Airflow installer and tmux sesssion manager"
readme = "README.md"
authors = [{ name = "Ross Turk", email = "ross@rossturk.com" }]
license = { file = "LICENSE" }
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
]
keywords = ["airflow", "tmux"]
dependencies = [
    "libtmux",
    "rich-click",
    "GitPython",
    "requests"
]
requires-python = ">=3.0"

[project.optional-dependencies]
dev = ["black", "bumpver"]

[project.urls]
Homepage = "https://github.com/rossturk/airflux"

[project.scripts]
airflux = "airflux.cli:main"

[tool.bumpver]
current_version = "1.0.0"
version_pattern = "MAJOR.MINOR.PATCH"
commit_message = "bump version {old_version} -> {new_version}"
commit = true
tag = true
push = true

[tool.bumpver.file_patterns]
"pyproject.toml" = [
    'current_version = "{version}"',
]
"airflux/cli.py" = [
    "{version}",
]
